refactor: move file upload logic to Pinia store
